Initial Commit
Current status as of git initialization:

Webserver working. Can take a frontend html textarea, send it to PHP
land, pass it on to R as an RMD file, render that in various formats,
and return them to the front end to be downloaded.

Editing interface partially done. Structure is sound, but about half the
options are not there.